requirements.txt
anyio==3.6.2
databases==0.7.0
fastapi==0.88.0
greenlet==2.0.1
idna==3.4
pydantic==1.10.2
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0

python-dotenv~=0.21.0
uvicorn~=0.20.0